[add]ルート保護を追加
ログイン時のみログインページ以外にアクセスできるように保護を追加しました。
また、リファクタリングも行っています。
