+++ skill-typing-front/app/components/game/gameResult.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Button from "./button";
+import Button from "../button";
 
 export default function GameResult({
   score,
